services/api/requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
psycopg2-binary==2.8.6
pytest==7.1.1